requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dateutil==2.8.2
cachetools==5.3.3 